refactor(video-library): extract public types to video-library.types.ts
First step of the video-library decomposition chantier (763 TS / 859 HTML / 1384
SCSS lines — well above the <400 soft cap). The 8 exported types
(`VideoItem`, `AddToTarget`, `VideoDeployState/Status`, `VideoContentStatus`,
`VideoOwnerType`, `SortField`, `SortDirection`) move to a dedicated types file
so they can be reused by upcoming sub-components (detail panel, preview modal,
filters) without a circular import through the component module.

The component re-exports the types (`export type { ... }`) so existing
importers (`video-manager`, `site-content-tab`) keep working without churn.
Saves ~31 lines in the main component. Zero behavioral change. All 709 smoke
tests in the video-library-related suites pass. `tsc --noEmit` clean.
